TASK: Change async interface and clean up
