+++ Classes/EventListener/AsynchronousEventListenerInterface.php
@@ -12,10 +12,25 @@ namespace Neos\Cqrs\EventListener;
  */
 
 /**
- * Marker interface for asynchronous event listeners
+ * Interface for asynchronous event listeners
  *
  * Asynchronous event listeners are not triggered directly when new events are published, but only when executing the CLI command projection:catchup
  */
 interface AsynchronousEventListenerInterface extends EventListenerInterface
 {
+    /**
+     * Returns the last seen sequence number of events which has been applied to the concrete event listener.
+     *
+     * @return int
+     */
+    public function getHighestAppliedSequenceNumber(): int;
+
+    /**
+     * Saves the $sequenceNumber as the last seen sequence number of events which have been applied to the concrete
+     * event listener.
+     *
+     * @param int $sequenceNumber
+     * @return void
+     */
+    public function saveHighestAppliedSequenceNumber(int $sequenceNumber);
 }
