+++ Classes/Projection/Doctrine/AbstractDoctrineProjector.php
@@ -11,9 +11,6 @@ namespace Neos\Cqrs\Projection\Doctrine;
  * source code.
  */
 
-use Doctrine\Common\Persistence\ObjectManager as DoctrineObjectManager;
-use Doctrine\ORM\EntityManager as DoctrineEntityManager;
-use Neos\Cqrs\EventListener\AppliedEventsAwareEventListener;
 use Neos\Cqrs\Exception;
 use Neos\Cqrs\Projection\AbstractBaseProjector;
 use TYPO3\Flow\Annotations as Flow;
@@ -23,27 +20,13 @@ use TYPO3\Flow\Annotations as Flow;
  *
  * @api
  */
-abstract class AbstractDoctrineProjector extends AbstractBaseProjector implements AppliedEventsAwareEventListener
+abstract class AbstractDoctrineProjector extends AbstractBaseProjector
 {
     /**
      * @Flow\Inject
      * @var DoctrineProjectionPersistenceManager
      */
     protected $projectionPersistenceManager;
-
-    /**
-     * @var DoctrineEntityManager
-     */
-    private $entityManager;
-
-    /**
-     * @param DoctrineObjectManager $entityManager
-     * @return void
-     */
-    public function injectEntityManager(DoctrineObjectManager $entityManager)
-    {
-        $this->entityManager = $entityManager;
-    }
 
     /**
      * Initialize the Read Model class name
@@ -135,51 +118,5 @@ abstract class AbstractDoctrineProjector extends AbstractBaseProjector implement
     public function isEmpty()
     {
         return $this->projectionPersistenceManager->count($this->getReadModelClassName()) === 0;
-    }
-
-    /**
-     * Returns the last seen sequence number of events which has been applied to the concrete event listener.
-     *
-     * @return int
-     */
-    public function getHighestAppliedSequenceNumber(): int
-    {
-        $projectorIdentifier = $this->renderProjectorIdentifier(get_class($this));
-        $projectionState = $this->entityManager->find(ProjectionState::class, $projectorIdentifier);
-        return ($projectionState instanceof ProjectionState ? $projectionState->highestAppliedSequenceNumber : 0);
-    }
-
-    /**
-     * Saves the $sequenceNumber as the last seen sequence number of events which have been applied to the concrete
-     * event listener.
-     *
-     * @param int $sequenceNumber
-     * @return void
-     */
-    public function saveSequenceNumber(int $sequenceNumber)
-    {
-        $projectorIdentifier = $this->renderProjectorIdentifier(get_class($this));
-        $projectionState = $this->entityManager->find(ProjectionState::class, $projectorIdentifier);
-        if ($projectionState === null) {
-            $projectionState = new ProjectionState();
-            $projectionState->projectorIdentifier = $projectorIdentifier;
-        }
-        $projectionState->highestAppliedSequenceNumber = $sequenceNumber;
-        $this->entityManager->persist($projectionState);
-    }
-
-    /**
-     * Renders a projector identifier which can be used as an id in the projection state
-     *
-     * @param string $className
-     * @return string
-     */
-    private function renderProjectorIdentifier(string $className): string
-    {
-        $identifier = strtolower(str_replace('\\', '_', $className));
-        if (strlen($identifier) > 255) {
-            $identifier = substr($identifier, 0, 255 - 6) . '_' . substr(sha1($identifier), 0, 5);
-        }
-        return $identifier;
     }
 }
