+++ Classes/Projection/Doctrine/ProjectionState.php
@@ -28,7 +28,6 @@ class ProjectionState
     public $projectorIdentifier;
 
     /**
-    /**
      * @var integer
      */
     public $highestAppliedSequenceNumber;
